rag-app/pyproject.toml
[project]
name = "rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.13",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-groq>=0.3.4",
    "pymupdf>=1.26.1",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=4.1.0",
]
